调用retry、timeout方法之后，限制只能调用awaitXxx方法
